lifestyle emoji and price model

+++ shouters/models.py
@@ -100,12 +100,12 @@ class Shouter(models.Model):
     ig_price_story_post_fc = models.IntegerField(null=True, blank=True)
     ig_price_story_post_ugc = models.IntegerField(null=True, blank=True)
 
-    ig_fb_price_story_fc = models.FloatField(null=True, blank=True)
-    ig_fb_price_story_ugc = models.FloatField(null=True, blank=True)
-    ig_fb_price_post_fc = models.FloatField(null=True, blank=True)
-    ig_fb_price_post_ugc = models.FloatField(null=True, blank=True)
-    ig_fb_price_story_post_fc = models.FloatField(null=True, blank=True)
-    ig_fb_price_story_post_ugc = models.FloatField(null=True, blank=True)
+    ig_fb_price_story_fc = models.IntegerField(null=True, blank=True)
+    ig_fb_price_story_ugc = models.IntegerField(null=True, blank=True)
+    ig_fb_price_post_fc = models.IntegerField(null=True, blank=True)
+    ig_fb_price_post_ugc = models.IntegerField(null=True, blank=True)
+    ig_fb_price_story_post_fc = models.IntegerField(null=True, blank=True)
+    ig_fb_price_story_post_ugc = models.IntegerField(null=True, blank=True)
 
     # # Work Selection
     is_check_ig = models.BooleanField(default=False)
